Integrates WebsocketApi feeds into the feeds package
