+++ numismatic/feeds/__init__.py
@@ -1,12 +1,14 @@
 from .base import Feed
+from .bitfinex import BitfinexFeed
 from .bravenewcoin import BraveNewCoinFeed
 from .cryptocompare import CryptoCompareFeed
+from .gdax import GDAXFeed
 from .luno import LunoFeed
-
-__all__ = ["Feed", "CryptoCompare", "Luno", "BraveNewCoin"]
-
 
 from ..libs.utils import make_get_subclasses, subclass_factory
 
 setattr(Feed, '_get_subclasses', make_get_subclasses('Feed'))
 setattr(Feed, 'factory', subclass_factory)
+
+
+__all__ = ["Feed"].extend(Feed._get_subclasses().keys())
