+++ numismatic/feeds/gdax.py
@@ -8,15 +8,33 @@ from streamz import Stream
 import attr
 import websockets
 
-from .base import Exchange
+from .base import Feed, WebsocketApi
 from ..libs.events import Heartbeat, Trade, LimitOrder, CancelOrder
 
 logger = logging.getLogger(__name__)
 
 
+class GDAXFeed(Feed):
+
+    def __init__(self, **kwargs):
+        self.rest_api = None
+        self.ws_api = GDAXWebsocketApi(**{a.name:kwargs[a.name] for a 
+                                          in attr.fields(GDAXWebsocketApi)
+                                          if a.name in kwargs})
+
+    def get_list(self):
+        raise NotImplemented()
+
+    def get_info(self, assets):
+        raise NotImplemented()
+ 
+    def get_prices(self, assets, currencies):
+        raise NotImplemented()       
+
+
 @attr.s
-class GDAXExchange(Exchange):
-    '''Websocket client for the GDAX Exchange
+class GDAXWebsocketApi(WebsocketApi):
+    '''Websocket client for the GDAX WebsocketApi
 
     This currently opens a separate socket for every symbol that we listen to.
     This could probably be handled by having just one socket.
@@ -118,7 +136,7 @@ if __name__=='__main__':
     output_stream = Stream()
     printer = output_stream.map(print)
 
-    bfx = GDAXExchange(output_stream=output_stream)
+    bfx = GDAXWebsocketApi(output_stream=output_stream)
     bfx_btc = bfx.listen('BTC-USD', 'ticker,heartbeat')
 
     loop = asyncio.get_event_loop()
